requirements.txt
certifi==2018.4.16
chardet==3.0.4
dataclasses==0.6
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.0.6
django-ckeditor==5.4.0
django-filter==1.1.0
django-heroku==0.3.1
django-jet==1.0.7
django-js-asset==1.1.0
django-taggit==0.22.2
djangorestframework==3.8.2
djongo==1.2.29
gunicorn==19.8.1
idna==2.6
Jinja2==2.9.6
Markdown==2.6.11
MarkupSafe==1.0
oauthlib==2.1.0
olefile==0.45.1
Pillow==4.2.1
psycopg2==2.7.5
PyJWT==1.6.4
pymongo==3.6.1
python3-openid==3.1.0
pytz==2018.4
requests==2.18.4
requests-oauthlib==1.0.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
sqlparse==0.2.4
urllib3==1.22
virtualenv==15.1.0
virtualenvwrapper-win==1.2.5
whitenoise==3.3.1
wincertstore==0.2
youtube-dl==2017.12.31
